+++ NEWS.md
@@ -1,3 +1,13 @@
+# hyd1d 0.5.4
+
+* fix summary.WaterLevelDataFrame() for all missing data
+* fix waterLevel() for all missing data in df.gauging_data
+* fix html figure styling of man pages
+* updated roxygen version
+* fix dbdriver selection
+* fix broken link to INFORM webpage
+* import validated gauging data for 2024
+
 # hyd1d 0.5.3
 
 * require R >= 4.1.0